refactor(app): removed redux persist
